Conflicts:
	Gemfile
	Gemfile.lock
	VERSION
	app/assets/javascripts/project_users_select.js.coffee
	app/assets/stylesheets/base/gl_variables.scss
	app/assets/stylesheets/pages/merge_requests.scss
	app/controllers/admin/users_controller.rb
	app/controllers/groups/group_members_controller.rb
	app/controllers/help_controller.rb
	app/controllers/profiles_controller.rb
	app/controllers/projects/wikis_controller.rb
	app/helpers/groups_helper.rb
	app/helpers/projects_helper.rb
	app/helpers/search_helper.rb
	app/helpers/selects_helper.rb
	app/mailers/emails/groups.rb
	app/mailers/emails/profile.rb
	app/models/application_setting.rb
	app/models/hooks/web_hook.rb
	app/models/note.rb
	app/models/project.rb
	app/services/files/create_service.rb
	app/services/notification_service.rb
	app/views/admin/application_settings/_form.html.haml
	app/views/admin/applications/_form.html.haml
	app/views/admin/broadcast_messages/index.html.haml
	app/views/admin/groups/show.html.haml
	app/views/admin/users/show.html.haml
	app/views/dashboard/_projects.html.haml
	app/views/dashboard/milestones/index.html.haml
	app/views/dashboard/projects/starred.html.haml
	app/views/devise/mailer/confirmation_instructions.html.erb
	app/views/devise/mailer/reset_password_instructions.html.erb
	app/views/devise/mailer/unlock_instructions.html.erb
	app/views/devise/passwords/edit.html.haml
	app/views/devise/registrations/edit.html.erb
	app/views/devise/shared/_signup_box.html.haml
	app/views/events/event/_push.html.haml
	app/views/groups/_projects.html.haml
	app/views/groups/group_members/_group_member.html.haml
	app/views/groups/group_members/_new_group_member.html.haml
	app/views/groups/milestones/index.html.haml
	app/views/groups/projects.html.haml
	app/views/layouts/_head_panel.html.haml
	app/views/layouts/nav/_admin.html.haml
	app/views/layouts/nav/_project.html.haml
	app/views/notify/project_was_moved_email.html.haml
	app/views/notify/repository_push_email.html.haml
	app/views/profiles/accounts/show.html.haml
	app/views/profiles/emails/index.html.haml
	app/views/profiles/history.html.haml
	app/views/profiles/keys/index.html.haml
	app/views/profiles/show.html.haml
	app/views/projects/_commit_button.html.haml
	app/views/projects/_dropdown.html.haml
	app/views/projects/_home_panel.html.haml
	app/views/projects/_issuable_form.html.haml
	app/views/projects/branches/_branch.html.haml
	app/views/projects/commit/_commit_box.html.haml
	app/views/projects/commits/_head.html.haml
	app/views/projects/deploy_keys/_deploy_key.html.haml
	app/views/projects/deploy_keys/index.html.haml
	app/views/projects/diffs/_file.html.haml
	app/views/projects/diffs/_stats.html.haml
	app/views/projects/edit.html.haml
	app/views/projects/issues/_discussion.html.haml
	app/views/projects/issues/_issue.html.haml
	app/views/projects/issues/_issue_context.html.haml
	app/views/projects/issues/index.html.haml
	app/views/projects/labels/_form.html.haml
	app/views/projects/merge_requests/_discussion.html.haml
	app/views/projects/merge_requests/_new_submit.html.haml
	app/views/projects/merge_requests/_show.html.haml
	app/views/projects/merge_requests/show/_context.html.haml
	app/views/projects/merge_requests/show/_participants.html.haml
	app/views/projects/merge_requests/show/_state_widget.html.haml
	app/views/projects/milestones/_milestone.html.haml
	app/views/projects/milestones/show.html.haml
	app/views/projects/new.html.haml
	app/views/projects/notes/_form.html.haml
	app/views/projects/project_members/_new_project_member.html.haml
	app/views/projects/project_members/_project_member.html.haml
	app/views/projects/protected_branches/index.html.haml
	app/views/projects/repositories/_download_archive.html.haml
	app/views/projects/services/_form.html.haml
	app/views/projects/show.html.haml
	app/views/projects/tags/_tag.html.haml
	app/views/shared/_issuable_filter.html.haml
	app/views/snippets/current_user_index.html.haml
	app/views/snippets/index.html.haml
	app/views/snippets/show.html.haml
	app/views/users/_profile.html.haml
	app/views/users/_projects.html.haml
	app/views/users/calendar.html.haml
	app/views/users/show.html.haml
	config/routes.rb
	db/schema.rb
	doc/markdown/markdown.md
	doc/permissions/permissions.md
	doc/public_access/public_access.md
	doc/raketasks/backup_restore.md
	doc/update/6.x-or-7.x-to-7.10.md
	doc/workflow/README.md
	lib/gitlab/markdown.rb

+++ app/models/repository.rb
@@ -62,24 +62,28 @@ class Repository
 
   def add_branch(branch_name, ref)
     cache.expire(:branch_names)
+    @branches = nil
 
     gitlab_shell.add_branch(path_with_namespace, branch_name, ref)
   end
 
   def add_tag(tag_name, ref, message = nil)
     cache.expire(:tag_names)
+    @tags = nil
 
     gitlab_shell.add_tag(path_with_namespace, tag_name, ref, message)
   end
 
   def rm_branch(branch_name)
     cache.expire(:branch_names)
+    @branches = nil
 
     gitlab_shell.rm_branch(path_with_namespace, branch_name)
   end
 
   def rm_tag(tag_name)
     cache.expire(:tag_names)
+    @tags = nil
 
     gitlab_shell.rm_tag(path_with_namespace, tag_name)
   end
@@ -122,7 +126,7 @@ class Repository
 
   def expire_cache
     %i(size branch_names tag_names commit_count graph_log
-       readme version contribution_guide).each do |key|
+       readme version contribution_guide changelog license).each do |key|
       cache.expire(key)
     end
   end
@@ -145,29 +149,17 @@ class Repository
     end
   end
 
-  def timestamps_by_user_log(user)
-    author_emails = '(' + user.all_emails.map{ |e| Regexp.escape(e) }.join('|') + ')'
-    args = %W(git log -E --author=#{author_emails} --since=#{(Date.today - 1.year).to_s} --branches --pretty=format:%cd --date=short)
-    dates = Gitlab::Popen.popen(args, path_to_repo).first.split("\n")
-
-    if dates.present?
-      dates
-    else
-      []
-    end
-  end
-
-  def commits_per_day_for_user(user)
-    timestamps_by_user_log(user).
-      group_by { |commit_date| commit_date }.
-      inject({}) do |hash, (timestamp_date, commits)|
-        hash[timestamp_date] = commits.count
-        hash
-      end
+  def lookup_cache
+    @lookup_cache ||= {}
   end
 
   def method_missing(m, *args, &block)
-    raw_repository.send(m, *args, &block)
+    if m == :lookup && !block_given?
+      lookup_cache[m] ||= {}
+      lookup_cache[m][args.join(":")] ||= raw_repository.send(m, *args, &block)
+    else
+      raw_repository.send(m, *args, &block)
+    end
   end
 
   def respond_to?(method)
@@ -197,16 +189,44 @@ class Repository
   end
 
   def contribution_guide
-    cache.fetch(:contribution_guide) { tree(:head).contribution_guide }
+    cache.fetch(:contribution_guide) do
+      tree(:head).blobs.find do |file|
+        file.contributing?
+      end
+    end
+  end
+
+  def changelog
+    cache.fetch(:changelog) do
+      tree(:head).blobs.find do |file|
+        file.name =~ /\A(changelog|history)/i
+      end
+    end
+  end
+
+  def license
+    cache.fetch(:license) do
+      tree(:head).blobs.find do |file|
+        file.name =~ /\Alicense/i
+      end
+    end
   end
 
   def head_commit
-    commit(self.root_ref)
+    @head_commit ||= commit(self.root_ref)
+  end
+
+  def head_tree
+    @head_tree ||= Tree.new(self, head_commit.sha, nil)
   end
 
   def tree(sha = :head, path = nil)
     if sha == :head
-      sha = head_commit.sha
+      if path.nil?
+        return head_tree
+      else
+        sha = head_commit.sha
+      end
     end
 
     Tree.new(self, sha, path)
@@ -247,6 +267,9 @@ class Repository
   # Remove archives older than 2 hours
   def clean_old_archives
     repository_downloads_path = Gitlab.config.gitlab.repository_downloads_path
+
+    return unless File.directory?(repository_downloads_path)
+
     Gitlab::Popen.popen(%W(find #{repository_downloads_path} -not -path #{repository_downloads_path} -mmin +120 -delete))
   end
 
@@ -332,6 +355,18 @@ class Repository
     else
       []
     end
+  end
+
+  def branches
+    @branches ||= raw_repository.branches
+  end
+
+  def tags
+    @tags ||= raw_repository.tags
+  end
+
+  def root_ref
+    @root_ref ||= raw_repository.root_ref
   end
 
   private
